Expose TypeDefs publicly, pass it around
Gross. The fact that semantic information affects parsing is going to
become even worse if we ever try to re-process changed files
incrementally.

Note also that TypeDefs used to contain a borrow from the input string,
but I decided I don't want to have to keep the input string around
forever. If we did, we'd have to keep each parsed input string in
cli::main alive until we're done parsing all of the files. I'd rather
not.