Do not track typedefs when parsing
This forces us to emit an explicit 'ambiguous' construct when parsing,
carry it though in lower, and finally choose between the two
possibilities in statics (where we have the information to do so).

The bad news is that error messages for ambiguous constructs are less
granular, applying to the whole simp instead of e.g. just the variable
if it has the wrong type. This would probably be fixable if we had
`NameId`s.

The good news is that we can now parse each file completely
independently of other files. This means we could parse before
determining inter-file dependencies, and even in parallel if we wanted.
(Right now we don't do either of those things, even after this change.)